[Nordigen] Copy raw version of payee to memo
Keep a copy of the payee before doing any transformations in the memo
field.

Ref #113
