+++ reader/nordigen/mapper_test.go
@@ -48,122 +48,3 @@ func TestParseAmount(t *testing.T) {
 		})
 	}
 }
-
-func TestPayeeStripNonAlphanumeric(t *testing.T) {
-	want := "Im just alphanumeric"
-	got := payeeStripNonAlphanumeric("Im just alphanumeric")
-	if want != got {
-		t.Fatalf("alphanumeric: %s != %s", want, got)
-	}
-
-	want = "你好世界"
-	got = payeeStripNonAlphanumeric("你好世界")
-	if want != got {
-		t.Fatalf("non-english: %s != %s", want, got)
-	}
-
-	want = "Im not j ust alphanumeric"
-	got = payeeStripNonAlphanumeric("Im! not j.ust alphanumeric42 69")
-	if want != got {
-		t.Fatalf("non-alphanumeric: %s != %s", want, got)
-	}
-}
-
-func TestStrip(t *testing.T) {
-	type args struct {
-		s      string
-		strips []string
-	}
-	tests := []struct {
-		name string
-		args args
-		want string
-	}{
-		{name: "single",
-			args: args{s: "Im not here", strips: []string{"not "}},
-			want: "Im here",
-		},
-		{name: "multiple",
-			args: args{s: "Im not really here", strips: []string{"not ", "really "}},
-			want: "Im here",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := strip(tt.args.s, tt.args.strips); got != tt.want {
-				t.Errorf("Payee.Strip() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestPayeeFinder(t *testing.T) {
-	type args struct {
-		t       nordigen.Transaction
-		sources PayeeSources
-	}
-	tests := []struct {
-		name string
-		args args
-		want string
-	}{
-		// First source that yields a result should be used
-		{
-			name: "name,unstructured,additional",
-			args: args{
-				t: nordigen.Transaction{
-					CreditorName:                      "",
-					RemittanceInformationUnstructured: "",
-					AdditionalInformation:             "baz",
-				},
-				sources: PayeeSources{
-					{Name},
-					{Unstructured},
-					{Additional},
-				},
-			},
-			want: "baz",
-		},
-		// The "+" operator should concat fields
-		{
-			name: "name+unstructured,additional",
-			args: args{
-				t: nordigen.Transaction{
-					CreditorName:                      "foo",
-					RemittanceInformationUnstructured: "bar",
-					AdditionalInformation:             "baz",
-				},
-				sources: PayeeSources{
-					{Name, Unstructured},
-					{Additional},
-				},
-			},
-			want: "foo bar",
-		},
-		// No sources yields a result
-		{
-			name: "empty",
-			args: args{
-				t: nordigen.Transaction{
-					CreditorName:                      "",
-					RemittanceInformationUnstructured: "",
-					AdditionalInformation:             "",
-				},
-				sources: PayeeSources{
-					{Name},
-					{Unstructured},
-					{Additional},
-				},
-			},
-			want: "",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := payeeFinder(tt.args.t, tt.args.sources)
-			if got != tt.want {
-				t.Errorf("%v, want %v", got, tt.want)
-			}
-		})
-	}
-}
