+++ reader/nordigen/nordigen_test.go
@@ -71,7 +71,7 @@ func TestToYnabber(t *testing.T) {
 				ID:      ynabber.ID("foobar"),
 				Date:    time.Date(2025, time.May, 15, 0, 0, 0, 0, time.UTC),
 				Payee:   "SCOR",
-				Memo:    "",
+				Memo:    "5345 SCOR",
 				Amount:  ynabber.Milliunits(-469640)},
 			},
 			wantErr: false,
@@ -179,7 +179,7 @@ func TestToYnabber(t *testing.T) {
 				ID:      ynabber.ID("foobar"),
 				Date:    time.Date(2023, time.February, 24, 0, 0, 0, 0, time.UTC),
 				Payee:   "PASCAL AS",
-				Memo:    "",
+				Memo:    "PASCAL AS",
 				Amount:  ynabber.Milliunits(10000)},
 			},
 			wantErr: false,
